Update Minefield
Added a new field to the class, numOfMines.  This represents the number of mines for the board.  This allows for the board to be of varying sizes in the future, though we are only doing a 9x9 board right now.  Also, the getter for the numOfMines can be used by the mineTracker JLabel to set the number of mines at the beginning of the game without hard coding it in.  This will allow us to modify the board sizes in the future.